adding redux-toolkit as dependency and updating usersAction to be written in new sllice syntax
