+++ package.json
@@ -7,6 +7,7 @@
     "@manaflair/redux-batch": "1.0.0",
     "@material-ui/core": "4.12.4",
     "@material-ui/icons": "4.11.3",
+    "@reduxjs/toolkit": "^1.8.6",
     "axios": "0.27.2",
     "connected-react-router": "6.9.2",
     "jsts": "2.9.0",
